refactor(App) : separating a function into a Util file based on the nature of the function
- string-utils : String related util functions
- number-utils : Number related util functions
