+++ src/util/number-utils.js
@@ -1,0 +1,3 @@
+export function get_sum_of_number(arr) {
+  return arr.reduce((acc, cur) => acc + cur, 0);
+}
